@@ -35,12 +35,13 @@
     "eslint-plugin-mocha": "^9.0.0",
     "esm": "^3.2.22",
     "fuzzer": "^0.2.1",
-    "gh-pages": "^2.0.1",
+    "gh-pages": "^3.2.3",
     "intl-pluralrules": "^1.0.0",
     "mocha": "^9.0.2",
     "prettyjson": "^1.2.1",
-    "rollup": "^1.9.3",
-    "typedoc": "^0.20.34",
-    "typescript": "^4.2.3"
+    "rollup": "^2.56.3",
+    "sinon": "^11.1.2",
+    "typedoc": "^0.22.3",
+    "typescript": "^4.4.3"
   }
 }
