@@ -3,9 +3,8 @@ const { resolve } = require("path");
 const { name } = require(resolve("package.json"));
 
 module.exports = {
-  entryPoints: ["src/index.ts"],
+  entryPoints: [resolve("src/index.ts")],
   hideGenerator: true,
   includeVersion: true,
-  //logger: "none",
-  out: name.replace("@fluent", "../html"),
+  out: resolve(name.replace("@fluent", "../html")),
 };
